Fix login with two-factor authentication
When your account has two-factor authentication, Twitter may use an
in-app browser to prompt for the code. Previously it was just a webview,
but that could not be used to authenticate with security keys (WebAuthn),
so they mostly use the in-app browser now. We don't want to redirect that
to Safari, or else it won't log in properly inside the app.